HOTFIX: Fixed carbon emission formula
We were treating the food_type.carbon_footprint value as kilograms and
just multiplying it by the food size. That value is actually in grams
and so we need to multiple by the average calorie size of the food type.

Also added seeds for average size of each food type.