HOTFIX: page error when user with no profile is on leaderboard
In some instance a user may not have a profile. If they end up on the
leaderboard, we try to get their name to display and page errors out
with can’t find first_name of nil.

Added a check to make sure each user on leaderboad has a profile before
showing.